Reorganized stack structure
