+++ python/route53-failover/cdk.json
@@ -65,6 +65,10 @@
     "@aws-cdk/aws-eks:nodegroupNameAttribute": true,
     "@aws-cdk/aws-ec2:ebsDefaultGp3Volume": true,
     "@aws-cdk/aws-ecs:removeDefaultDeploymentAlarm": true,
-    "@aws-cdk/custom-resources:logApiResponseDataPropertyTrueDefault": false
+    "@aws-cdk/custom-resources:logApiResponseDataPropertyTrueDefault": false,
+    "domain": "junsjang.link",
+    "email": "junsjang@amazon.com",
+    "primaryRegion": "ap-northeast-1",
+    "secondaryRegion": "ap-northeast-2"
   }
 }
